// rickandmorty/src/types/api/character.d.ts
export interface ICharactersResponse {
    data: ICharacter[]
}


export interface ICharacter {
    id: string,
    name: string,
    status: string,
    species: string,
    type: string,
    gender: string,
    origin: IOrigin,
    location: ILocation,
    image: string,
    episode: string,
    url: string,
    created: string
}


interface IOrigin {
    name: string,
    url: string
}

interface ILocation {
    name: string,
    url: string
}